win: mark recent Studio Drivers
Nvidia is recently releasing same driver version as both Game Ready and
Studio. This might have some changes in the GPU drivers, but the DLLs
for NVENC and NVFBC are exactly the same (byte-by-byte) and same patches
apply on both variants.

Update the README to indicate this commonality in the patches to avoid
confusion.

Signed-off-by: Jai Luthra <me@jailuthra.in>