linux: Add `patch-vk` which patches `libnvidia-eglcore.so` for vulkan
This bypasses the limit for vulkan drivers.

Whenever an encoder session is created, it will call a function that
does "authorization", and then before it exit or when the session is
closed, it will "release" it.

This happens at the function 0xD7E780 (address in version 555.58.02).

The function signature is like this
```C
// This is C++, but for `C` translation it will be something like this
bool auth_release(void* this, bool is_auth);
```

It will take either `is_auth==true` if in `auth` or `false` when
releasing.

It will return either `true` for success or `false` for failure.

What we changed here is that we replaced this function implementation
with `return true;`.

And it seems to be working as expected, and we can bypass the limit by
not even registering and checking for it.

Part of this commit, I checked the other surrounding versions, and they
all seem to use the exact same function which is good for us.

I actually didn't test any version beside 555.58.02, so it could need
some testing.

Signed-off-by: Amjad Alsharafi <amjadsharafi10@gmail.com>